+++ TrelloDotNet/TrelloDotNet/Model/Member.cs
@@ -1,5 +1,4 @@
-﻿using System.Collections.Generic;
-using System.Diagnostics;
+﻿using System.Diagnostics;
 using System.Text.Json.Serialization;
 
 namespace TrelloDotNet.Model